# docs/swagger.yaml
basePath: /
definitions:
  domain.Receipt:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      image_url:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.ReceiptItem'
        type: array
      merchant:
        type: string
      subtotal:
        type: number
      tax:
        type: number
      total:
        type: number
      updated_at:
        type: string
    type: object
  domain.ReceiptItem:
    properties:
      category:
        type: string
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      qty:
        type: integer
    type: object
...
  model.CategorySummary:
    properties:
      amount:
        type: string
      category:
        type: string
      percentage:
        type: number
    type: object
  model.DashboardSummaryResponse:
    properties:
      averageSpend:
        type: string
      receiptCount:
        type: integer
      topCategories:
        items:
          $ref: '#/definitions/model.CategorySummary'
        type: array
      topMerchants:
        items:
          $ref: '#/definitions/model.MerchantSummary'
        type: array
      totalSpend:
        type: string
    type: object
  model.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/model.ErrorDetail'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
  model.MerchantSummary:
    properties:
      amount:
        type: string
      merchant:
        type: string
      percentage:
        type: number
    type: object
  model.PaginationResponse:
    properties:
      currentPage:
        type: integer
      limit:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
  model.ReceiptItemResponse:
    properties:
      category:
        type: string
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: string
      qty:
        type: integer
    type: object
  model.ReceiptResponse:
    properties:
      createdAt:
        type: string
      date:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/model.ReceiptItemResponse'
        type: array
      merchant:
        type: string
      subtotal:
        type: string
      tax:
        type: string
      total:
        type: string
      updatedAt:
        type: string
    type: object
  model.ReceiptsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ReceiptResponse'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationResponse'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for scanning and managing receipts with AI-powered data extraction
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Receipt Scanner API
  version: "1.0"
paths:
...
  /v1/dashboard/summary:
    get:
      consumes:
      - application/json
      description: Get summary statistics for the dashboard
      parameters:
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard summary
          schema:
            $ref: '#/definitions/model.DashboardSummaryResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get dashboard summary
      tags:
      - dashboard
  /v1/receipts:
    get:
      consumes:
      - application/json
      description: Get a paginated list of receipts with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - description: Merchant name filter
        in: query
        name: merchant
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of receipts
          schema:
            $ref: '#/definitions/model.ReceiptsListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: List all receipts
      tags:
      - receipts
    post:
      consumes:
      - application/json
      description: Create a new receipt with manual data entry
      parameters:
      - description: Receipt data
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/domain.Receipt'
      produces:
      - application/json
      responses:
        "201":
          description: Receipt created successfully
          schema:
            $ref: '#/definitions/model.ReceiptResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create a new receipt
      tags:
      - receipts
  /v1/receipts/{receiptId}:
    delete:
      consumes:
      - application/json
      description: Delete a receipt by ID
      parameters:
      - description: Receipt ID
        in: path
        name: receiptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Receipt deleted successfully
        "400":
          description: Invalid receipt ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Receipt not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Delete a receipt
      tags:
      - receipts
    get:
      consumes:
      - application/json
      description: Retrieve a specific receipt by its ID
      parameters:
      - description: Receipt ID
        in: path
        name: receiptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Receipt details
          schema:
            $ref: '#/definitions/model.ReceiptResponse'
        "400":
          description: Invalid receipt ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Receipt not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get a receipt by ID
      tags:
      - receipts
    put:
      consumes:
      - application/json
      description: Update an existing receipt by ID
      parameters:
      - description: Receipt ID
        in: path
        name: receiptId
        required: true
        type: string
      - description: Updated receipt data
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/domain.Receipt'
      produces:
      - application/json
      responses:
        "200":
          description: Receipt updated successfully
          schema:
            $ref: '#/definitions/model.ReceiptResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Receipt not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Update a receipt
      tags:
      - receipts
  /v1/receipts/scan:
    post:
      consumes:
      - multipart/form-data
      description: Upload and process a receipt image to extract data using AI
      parameters:
      - description: Receipt image file
        in: formData
        name: receiptImage
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully scanned receipt
          schema:
            $ref: '#/definitions/model.ReceiptResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unable to extract data
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Scan a receipt image
      tags:
      - receipts
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"